// src/frontend/src/api/device.d.ts
export interface components {
    schemas: {
        DeviceCredentialResponse: {
            device: components["schemas"]["DeviceResponse"];
            credential: string;
        };
        DevicePage: {
            items: components["schemas"]["DeviceResponse"][];
            /** Format: uuid */
            nextCursor: null | string;
        };
        DeviceResponse: {
            /** Format: uuid */
            id: string;
            name: string;
            kind: string;
            permission: string;
            isOwner: boolean;
            /** Format: date-time */
            createdAt: string;
            /** Format: int32 */
            version: number;
        };
        ProblemDetails: {
            type?: null | string;
            title?: null | string;
            /** Format: int32 */
            status?: null | number;
            detail?: null | string;
            instance?: null | string;
        };
        RegisterDeviceRequest: {
            name: string;
            kind: string;
        };
        UpdateDeviceRequest: {
            name: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}